ibge-pipeline/apps/load-app/requirements.txt
pandas==2.*
kafka-python==2.*
opentelemetry-api==1.*
opentelemetry-sdk==1.*
opentelemetry-exporter-otlp==1.*
opentelemetry-instrumentation-confluent-kafka
